+++ src/components/formcomponents/open-choice/__tests__/openchoice-dropdown-view-spec.tsx
@@ -278,6 +278,27 @@ describe('Dropdown-view - choice', () => {
           await userEvent.tab();
           expect(queryByText(resources.formRequiredErrorMessage)).not.toBeInTheDocument();
         });
+        it('Should get required error on readOnly if noe value', async () => {
+          const questionnaire: Questionnaire = {
+            ...q,
+            item: q.item?.map(x => ({
+              ...x,
+              readOnly: true,
+              required: true,
+              code: [
+                {
+                  code: 'ValidateReadOnly',
+                  display: 'Valider skrivebeskyttet felt',
+                  system: 'http://helsenorge.no/fhir/CodeSystem/ValidationOptions',
+                },
+              ],
+            })),
+          };
+          const { getByText } = createWrapper(questionnaire);
+          await submitForm();
+
+          expect(getByText(resources.formRequiredErrorMessage)).toBeInTheDocument();
+        });
       });
     });
   });