+++ src/components/formcomponents/open-choice/dropdown-view.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { FieldValues, useFormContext } from 'react-hook-form';
+import { FieldValues, RegisterOptions, useFormContext } from 'react-hook-form';
 import styles from '../common-styles.module.css';
 import { Options } from '@/types/formTypes/radioGroupOptions';
 
@@ -44,26 +44,35 @@ const DropdownView = (props: Props): JSX.Element | null => {
 
   const { resources } = useExternalRenderContext();
   const answer = useGetAnswer(linkId, path);
-  const onChange = (e: React.ChangeEvent<HTMLSelectElement>): void => {
-    handleChange(e.target.value);
-  };
-  const { onChange: handleFormChange, ...rest } = register(idWithLinkIdAndItemIndex, {
-    required: required({ item, resources }),
-    shouldUnregister: true,
-  });
-  if (!options) {
-    return null;
-  }
-  const getWith = (options: Array<Options>): number => {
+
+  const getWith = (options: Array<Options> | undefined): number => {
     const maxCharacters = options?.reduce((acc, option) => (option.label.length > acc ? option.label.length : acc), 0);
     const placeholderLength = resources?.selectDefaultPlaceholder ? resources.selectDefaultPlaceholder.length : 0;
     let width = maxCharacters ? (maxCharacters > 40 ? 40 : maxCharacters) : 25;
     return (width = placeholderLength > width ? placeholderLength : width);
   };
 
+  const validationRules: RegisterOptions<FieldValues, string> | undefined = {
+    required: required({ item, resources }),
+    shouldUnregister: true,
+  };
+
+  const onChange = (e: React.ChangeEvent<HTMLSelectElement>): void => {
+    handleChange(e.target.value);
+  };
+  const { onChange: handleFormChange, ...rest } = register(idWithLinkIdAndItemIndex, pdf ? undefined : validationRules);
+
   if (pdf || isReadOnly(item)) {
     return (
-      <ReadOnly pdf={pdf} id={id} item={item} pdfValue={pdfValue} errors={error}>
+      <ReadOnly
+        pdf={pdf}
+        id={id}
+        idWithLinkIdAndItemIndex={idWithLinkIdAndItemIndex}
+        item={item}
+        value={selected}
+        pdfValue={pdfValue}
+        errors={error}
+      >
         {children}
       </ReadOnly>
     );
@@ -96,7 +105,7 @@ const DropdownView = (props: Props): JSX.Element | null => {
           value={selected?.[0] || ''}
         >
           <option value={undefined}>{resources?.selectDefaultPlaceholder || ''}</option>
-          {options.map(option => (
+          {options?.map(option => (
             <option key={getId(id) + option.label} value={option.type}>
               {option.label}
             </option>