+++ CHANGES.md
@@ -1,3 +1,9 @@
+## 16.0.0-beta.42
+- fix redux error calling updateSkjemaDefinisjonAction
+
+## 16.0.0-beta.42
+- Made validation possible for readOnly fields
+
 ## 16.0.0-beta.41
 - Added redux toolkit